Add docs input `Email::class`.
