+++ docs/README.md
@@ -57,6 +57,7 @@ This is particularly useful for creating forms and other interactive elements on
 - [date](/docs/form/input/Date.md)
 - [datetime](/docs/form/input/Datetime.md)
 - [datetime-local](/docs/form/input/DatetimeLocal.md)
+- [email](/docs/form/input/Email.md)
 - [radio](/docs/form/input/Radio.md)
 - [radio-list](/docs/form/input/RadioList.md)
 - [text](/docs/form/input/Text.md)
